chore: refactor the matchup calculation to dedicated job
